FileReader: is moved into separated file
