+++ modules/affile/Makefile.am
@@ -5,6 +5,8 @@ modules_affile_libaffile_la_SOURCES	=			\
 	modules/affile/logproto-file-writer.h			\
 	modules/affile/poll-file-changes.c			\
 	modules/affile/poll-file-changes.h			\
+	modules/affile/file-reader.c				\
+	modules/affile/file-reader.h				\
 	modules/affile/affile-common.c				\
 	modules/affile/affile-common.h				\
 	modules/affile/affile-source.c				\
